+++ components/home/popular/Popularjobs.jsx
@@ -10,12 +10,10 @@ import useFetch from '../../../hooks/useFetch';
 const Popularjobs = () => {
   const router = useRouter();
 
-  const {data, isLoading, error } = useFetch('search', {
-    query:'React Developer',
-    num_pages:1
+  const {data, isLoading, error } = useFetch("search", {
+    query:"React Developer",
+    num_pages:"1"
   });
-
-  console.log(data);
 
   return (
     <View style={styles.container}>