+++ hooks/useFetch.js
@@ -1,15 +1,15 @@
 import { useEffect, useState } from "react";
+import axios from "axios";
+
 
 const useFetch = (endpoint, query) => {
     const [data, setData] = useState([]);
     const [isLoading, setIsLoading] = useState(false);
     const [error, setError] = useState(null);
 
-    const axios = require('axios');
-
     const options = {
         method: 'GET',
-        url: `https://jsearch.p.rapidapi.com/search${endpoint}`,
+        url: `https://jsearch.p.rapidapi.com/${endpoint}`,
         headers: {
             'X-RapidAPI-Key': '2380a6b96bmshe73b6501b57c207p15d141jsnd3edf7c68147',
             'X-RapidAPI-Host': 'jsearch.p.rapidapi.com'
@@ -24,17 +24,17 @@ const useFetch = (endpoint, query) => {
             setData(response.data.data);
             setIsLoading(false);
         } catch (error) {
+            console.log(error);
             setError(error);
-            console.error(error);
         } finally {
-            isLoading(false);
+            setIsLoading(false);
         }
-    }
+    };
 
     const refetchData = () => {
         setIsLoading(true);
         fetchData();
-    }
+    };
 
     useEffect(() => {
         fetchData();